requirements.txt
PyGithub
requests
openai
